Add environment build switches
