+++ ts/frontend/src/environments/env-static.ai4realnet.ts
@@ -1,0 +1,9 @@
+export const envStatic = {
+  content: {
+    title: 'AI4REALNET Validation Campaign Hub',
+    home: {
+      lead: 'Welcome to the validation campaign hub. The validation campaign hub provides an overview of the evaluation objectives and their respective KPI of the [AI4REALNET project] (https://ai4realnet.eu/). Through this hub, project partners can evaluate their solutions for specific KPI. The corresponding procedures follow either a fully automated evaluation or for KPI requiring human-machine interactions either semi-automated or offline. To make a submission choose the evaluation objective and subsequently the KPI.',
+      benchmarksHeading: 'AI4REALNET Validation Campaign',
+    },
+  },
+}
